Divided the games shelf to component, hook and api
