+++ frontend/src/games/api/games.js
@@ -11,9 +11,19 @@ const BASE_URL = process.env.REACT_APP_BASE_URL;
  */
 export const fetchShelvesConfigs = async () => {
     const res = await axios.get(`${BASE_URL}/games/row-config`);
-    console.log(res.data);
     return res.data;
 };
+
+/**
+ * Fetches games for a specific game shelf
+ * @param fetchUrl shelf related pathing in URL
+ * @param params extra params (e.g. limit)
+ * @returns {Promise<any>} list of game objects
+ */
+export const fetchGamesByShelf = async (fetchUrl, params) => {
+    const res = await axios.get(`${BASE_URL}${fetchUrl}`, {params});
+    return res.data;
+}
 
 /**
  * Get all available ga,es from the backend
@@ -22,4 +32,4 @@ export const fetchShelvesConfigs = async () => {
 export const fetchGames = async () => {
     const response = await axios.get(`${BASE_URL}/games`);
     return response.data;
-}
+};
